Change version of parameterized models to 4.4
This commit modifies all the java docs and concerned files in previous
commits in this pull request.
