+++ subprojects/tooling-api/src/crossVersionTest/groovy/org/gradle/integtests/tooling/r44/CustomModel2.java
@@ -14,20 +14,8 @@
  * limitations under the License.
  */
 
-package org.gradle.integtests.tooling.r43;
+package org.gradle.integtests.tooling.r44;
 
-import org.gradle.api.Action;
-import org.gradle.tooling.BuildAction;
-import org.gradle.tooling.BuildController;
-
-public class ParameterAction implements BuildAction<CustomModel> {
-    @Override
-    public CustomModel execute(BuildController controller) {
-        return controller.getModel(CustomModel.class, CustomParameter.class, new Action<CustomParameter>() {
-            @Override
-            public void execute(CustomParameter customParameter) {
-                customParameter.setValue("myParameter");
-            }
-        });
-    }
+public interface CustomModel2 {
+    String getValue();
 }
