+++ subprojects/tooling-api/src/crossVersionTest/groovy/org/gradle/integtests/tooling/r44/MultipleParametersAction.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.integtests.tooling.r43;
+package org.gradle.integtests.tooling.r44;
 
 import org.gradle.api.Action;
 import org.gradle.tooling.BuildAction;
